+++ src/Sulu/Bundle/TagBundle/Admin/SuluTagAdmin.php
@@ -13,13 +13,12 @@ namespace Sulu\Bundle\TagBundle\Admin;
 use Sulu\Bundle\AdminBundle\Admin\Admin;
 use Sulu\Bundle\AdminBundle\Navigation\Navigation;
 use Sulu\Bundle\AdminBundle\Navigation\NavigationItem;
-use Sulu\Bundle\SecurityBundle\Permission\SecurityChecker;
-use Sulu\Bundle\SecurityBundle\Permission\SecurityCheckerInterface;
+use Sulu\Component\Security\Authorization\SecurityCheckerInterface;
 
 class SuluTagAdmin extends Admin
 {
     /**
-     * @var SecurityChecker
+     * @var SecurityCheckerInterface
      */
     private $securityChecker;